Trait promptly::Promptable

pub trait Promptable: Sized {
    fn prompt<S: AsRef<str>>(msg: S) -> Result<Self, ReadlineError>;

    fn prompt_opt<S: AsRef<str>>(msg: S) -> Result<Option<Self>, ReadlineError>;

    fn prompt_default<S: AsRef<str>>(
        msg: S, 
        default: Self
    ) -> Result<Self, ReadlineError>;
}

A trait for convenient, opinionated prompting

Required methods

fn prompt<S: AsRef<str>>(msg: S) -> Result<Self, ReadlineError>

Prompts for a value. Re-prompts on invalid and empty input.

fn prompt_opt<S: AsRef<str>>(msg: S) -> Result<Option<Self>, ReadlineError>

Prompts for a value, returning None for empty input. Re-prompts on invalid input.

fn prompt_default<S: AsRef<str>>(
msg: S,
default: Self
) -> Result<Self, ReadlineError>

Prompts for a value with a default value for empty input. Re-prompts on invalid input.

The default value will be mentioned in the prompt message

Loading content...